requirements.txt
Django==5.2.3
Babel==2.17.0
boto3==1.38.35
botocore==1.38.35
django-storages==1.14.6
defusedxml==0.7.1
django-allauth==65.9.0
PyJWT==2.10.1
python-openid==2.2.5
python3-openid==3.2.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
oauth2==1.9.0.post1
django-extensions==4.1
model-bakery==1.20.5
django-bootstrap3==25.1
django-countries==7.6.1
django-localflavor==5.0
django-phonenumber-field==8.1.0
phonenumberslite==9.0.7
Pillow==11.2.1
ipython==9.3.0
ipython-genutils==0.2.0
httplib2==0.22.0
requests==2.32.4
sendgrid==6.12.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
shortuuid==1.0.13
jmespath==1.0.1
docutils==0.21.2
pexpect==4.9.0
ptyprocess==0.7.0
decorator==5.2.1
simplegeneric==0.8.1
traitlets==5.14.3
pickleshare==0.7.5
smtpapi==0.4.12
python-http-client==3.3.7
django-stubs==5.2.0
django-stubs-ext==5.2.0
types-requests==2.32.4.20250611